refactor(p2): extract ChatListNormalSection into sibling file (stage 4 of 4)
Moves the 1021-line ChatListNormal composable (parent L322-1342 of
post-stage-3 ChatList.kt) into the new sibling ChatListNormalSection.kt
in the same package. This is the big timeline composable: state-machine
follow logic, ime auto-scroll wiring, lazy column rendering, virtual /
non-virtual message branches, selection toolbar, export sheet, message
jumper, suggestion strip overlay.

Byte-equal: body and signature unchanged. Inline FQNs preserved
verbatim (me.rerere.rikkahub.RouteActivity at L382, java.util.Locale
in onTranslate signature). Same-package neighbours (ChatJankProbe,
rememberConversationSizeInfo / ConversationSizeWarningDialog,
ChatExportSheet) need no imports.

Visibility widenings (1 sibling p→i): ChatListNormal — called by
ChatList entry that stays in coord.

Visibility widenings (15 coord p→i, for the sibling to reach the
file-private shared shapes still living in coord):
SCROLL_TAG, LoadingIndicatorKey, HistoryLoadingItemKey,
ScrollBottomKey, TimelineHorizontalPadding, SendTransitionSlideDistance,
TimelineHistoryLoadSignal (data class), TimelineScrollAnchor (data class),
summaryPreviewOf, TimelineTopPadding, TimelineBottomSafetyPadding,
TimelineItemSpacing, TimelineMessageInnerSpacing,
TimelineSelectionToolbarOffset, TimelineFollowMode (enum class).

Coordinator strip: 1342 -> 229 (-1113 with import cleanup).
Imports trimmed: 90 orphans — the full ChatListNormal-consumed surface
across Compose foundation/runtime/animation/material3/lazy + project
data/services + kotlinx coroutines + kotlin math/util + HugeIcons.
Three of the 90 (offset, size, width) were technically orphan since
Stage 2/3 but missed at those stages — pre-cut grep pattern lesson:
infix Modifier extensions whose simple name collides with a property
name (data class field "offset", DSL scope property "size") evade
the standard "import line + body ref" diff heuristic. Cleaned here.

Pre-existing dead imports preserved verbatim:
CircleShape, getValue, OutlinedTextField, Search01, setValue,
withTimeoutOrNull. Each verified parent=1 (only import line, no body
ref) at merge base 307ee8a5.

Gates:
- :app:compileDebugKotlin BUILD SUCCESSFUL
- :app:testDebugUnitTest — 3 failed / 467 (baseline preserved:
  GenerativeUiPlannerTest×2 + DefaultProvidersTest×1)
- :app:assembleDebug :app:assembleNotion :app:assembleRefactortest —
  all BUILD SUCCESSFUL (3-variant final-stage gate per playbook)
